# feat(perf): optimize encoder throughput.