# Context

* Add benchmarks with different output sizes
* Replace `bytes.Buffer` with a pre-allocated `strings.Builder`
* Relocate panic recovery function in the public function to reduce `defer` overhead in the recursive function
* Remove Regex usage

## Results

```
❯ benchstat old.txt new.txt
name                                old time/op    new time/op    delta
EncodeCanonical/input_size_146-10     19.2µs ± 1%     4.4µs ± 1%  -77.02%  (p=0.008 n=5+5)
EncodeCanonical/input_size_2713-10     244µs ± 1%      60µs ± 1%  -75.53%  (p=0.008 n=5+5)
EncodeCanonical/input_size_6412-10     530µs ± 5%     125µs ± 0%  -76.51%  (p=0.008 n=5+5)

name                                old alloc/op   new alloc/op   delta
EncodeCanonical/input_size_146-10     22.7kB ± 0%     3.3kB ± 0%  -85.37%  (p=0.008 n=5+5)
EncodeCanonical/input_size_2713-10     253kB ± 0%      42kB ± 0%  -83.34%  (p=0.008 n=5+5)
EncodeCanonical/input_size_6412-10     513kB ± 0%      85kB ± 0%  -83.39%  (p=0.008 n=5+5)

name                                old allocs/op  new allocs/op  delta
EncodeCanonical/input_size_146-10        331 ± 0%        87 ± 0%  -73.72%  (p=0.008 n=5+5)
EncodeCanonical/input_size_2713-10     3.35k ± 0%     0.79k ± 0%  -76.26%  (p=0.008 n=5+5)
EncodeCanonical/input_size_6412-10     6.64k ± 0%     1.58k ± 0%  -76.23%  (p=0.008 n=5+5)
```

# Reference(s)

- fix https://github.com/secure-systems-lab/go-securesystemslib/issues/24

Signed-off-by: Thibault Normand <me@zenithar.org>