@adityasaky 2022-06-20
This looks good to me but as Trishank pointed out, we shouldn't accidentally break the dependents. I'm merging it to a non-main branch now to allow us to test it in in-toto-golang / go-tuf.