@adityasaky 2022-06-20
cc @shibumi for in-toto-golang specific expertise. Using this version of go-sslib doesn't break any of in-toto's unit tests but that's of course never the full story. Does anything here worry you? :smile: