@mnm678 2022-11-10
I ran the go-tuf tests using the `go-securesystemslib` branch locally and everything passed. There might be something not covered by the tests, but I think this features is safe to merge from the go-tuf side.