@adityasaky 2022-11-10
Also looks good wrt tests on in-toto-golang. I think this should be okay to merge into main and patch any issues we encounter later.